Typescript/src/index.ts
import { Clock } from "./clock";

type Size = "small" | "big";

export type WeekDay =
  | "Monday"
  | "Tuesday"
  | "Wednesday"
  | "Thursday"
  | "Friday"
  | "Saturday"
  | "Sunday";

export interface Calendar {
  day: WeekDay;
  date: number;
  month: number;
  year: number;
}

type OrderDate = Calendar;
type DeliveryDate = Calendar;

type OrderTime = "Morning" | "Afternoon";

interface OrderOptions {
  orderDate: OrderDate;
  size: Size;
  isMorningOrder?: boolean;
  hasCustomFrosting?: boolean;
}

export function order({
  orderDate,
  size,
  hasCustomFrosting,
  isMorningOrder,
}: OrderOptions): DeliveryDate {
  const clock = new Clock(orderDate);
  let leadTime = size === "small" ? 2 : 3;
  if (isMorningOrder) {
    leadTime -= 1;
  }
  if (hasCustomFrosting) {
    leadTime += 1;
  }
  const deliveryDate = clock.add(leadTime);
  return deliveryDate.toCalendar();
}
